最新

+++ Project/Project/Source/SceneContains/Elements/AnimationFont/AnimationFont.cpp
@@ -3,8 +3,8 @@
 #include "../../../Base/GameManagerContains/IGameManager.h"
 #include <GSvector2.h>
 
-AnimationFont::AnimationFont(IWorld * world, const GSvector2 & position, const IGameManagerPtr & gameManager, const std::string texName, const int updateFrame)
-	: Actor(world, ActorName::UI_Sprite, position, gameManager, std::make_shared<AnimationTexture>(texName, gameManager->GetDrawManager(), DrawOrder::UI, 50, updateFrame)) {
+AnimationFont::AnimationFont(IWorld * world, const GSvector2 & position, const IGameManagerPtr & gameManager, const std::string texName, const float updateFrame)
+	: Actor(world, ActorName::UI_Sprite, position, gameManager, std::make_shared<AnimationTexture>(texName, gameManager->GetDrawManager(), DrawOrder::UI, 50.0f, updateFrame)) {
 }
 
 AnimationFont::~AnimationFont() {
